some more stuff

+++ 04.01.NetworkArchitecture/app/src/main/java/de/mobilecomputing/exercise3/networkarchitecture/reqresapi/ReqResApi.java
@@ -12,5 +12,5 @@ public interface ReqResApi {
     Call<List<User>> getUsers();
 
     @GET("api/users/{id}")
-    Call<User> getUser(@Path("id") int userId);
+    Call<User> getUser(@Path("id") long userId);
 }
